backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
pymongo==4.6.1
Werkzeug==3.0.1
google-generativeai==0.8.3
Pillow==10.1.0
python-dotenv==1.0.0
markdown==3.5.1
cloudinary
requests
